+++ sdk/ai/azure-ai-projects/azure/ai/projects/prompts/__init__.py
@@ -1,0 +1,15 @@
+# ------------------------------------
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+# ------------------------------------
+# pylint: disable=unused-import
+try:
+    import prompty  # pylint: disable=unused-import
+except ImportError:
+    raise ImportError(
+        "The 'prompty' package is required to use the 'azure.ai.projects.prompts' module. "
+        "Please install it by running 'pip install prompty'."
+    )
+
+from ._patch import patch_sdk as _patch_sdk, PromptTemplate
+_patch_sdk()